Upgrade Gradle, AGP, Kotlin, and dependencies to latest stable
Migrates to AGP 9's built-in Kotlin support (removes the separate
org.jetbrains.kotlin.android plugin), bumps Gradle to 9.6.1, Kotlin to
2.4.0, and compileSdk/targetSdk to 36. Also updates androidx/gms/suncalc
dependencies to their latest versions compatible with this toolchain,
and removes a redundant Float conversion flagged by the new compiler.

+++ app/src/main/java/com/atkinson/sunsetter/SunsetComplicationDataSourceService.kt
@@ -12,7 +12,7 @@ class SunsetComplicationDataSourceService : SuspendingComplicationDataSourceServ
     override suspend fun onComplicationRequest(request: ComplicationRequest): ComplicationData {
         val sunTimes = SunTimeRepository(this).getSunTimesForToday()
         return buildComplicationData(
-            value = (sunTimes.sunsetMinute - HALF_DAY).toFloat(),
+            value = sunTimes.sunsetMinute - HALF_DAY,
             text = minuteOfDayToTimeString(sunTimes.sunsetMinute)
         )
     }
